+++ CHANGELOG.md
@@ -6,6 +6,18 @@ The format is based on [Keep a Changelog][], and this project adheres to
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Unreleased]: https://github.com/yourbase/yb/compare/v0.4.0...HEAD
 
+## [Unreleased][]
+
+### Changed
+
+-  Attempting to use an unknown container in `{{.Container.IP}}` substitutions
+   will now cause a build failure rather than silently expanding to the
+   empty string.
+
+### Fixed
+
+-  Fixed the `{{.Container.IP}}` regression introduced in v0.4.0.
+
 ## [0.4.0][] - 2020-10-12
 
 Version 0.4 removes some broken or ill-conceived functionality from yb and